+++ src/faker.ts
@@ -1,10 +1,7 @@
 import "reflect-metadata";
 import * as dotenv from "dotenv";
-import { faker } from "@faker-js/faker";
-import { generateHash } from "./utils";
-import { DatabaseUser } from "./types";
 import { AppDataSource } from "./data-source";
-import { UserRepository } from "./data-source";
+import { fillDatabase } from "./utils";
 
 const startDB = async () => {
   dotenv.config({ path: "./.env" });
@@ -16,31 +13,6 @@ const initializeData = async () => {
     url: process.env.DB_URL,
   });
   await AppDataSource.initialize();
-};
-
-const fillDatabase = async (numberOfUsers: number) => {
-  const userArray: DatabaseUser[] = [];
-  for (let counter = 0; counter < numberOfUsers; counter++) {
-    userArray.push(generateRandomUser());
-  }
-  await UserRepository.insert(userArray);
-};
-
-const generateRandomUser = () => {
-  const newUser: DatabaseUser = {
-    name: faker.person.fullName(),
-    email: faker.internet.email(),
-    password: generateHash(
-      faker.internet.password({
-        length: 10,
-        pattern: /[A-Za-z\d]/,
-        prefix: "aA1",
-      })
-    ),
-    profession: faker.person.jobDescriptor(),
-    dateOfBirth: faker.date.birthdate(),
-  };
-  return newUser;
 };
 
 const fill = async (amount: number) => {
